refactor: align REST v2 struct array implementation
